LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focus-selection-expected.txt

button
foo

FAIL div has contenteditable=false assert_equals: Should have 1 selected range after focusing input type = submit expected 1 but got 0
FAIL div has contenteditable=true assert_equals: Should have 1 selected range after focusing input type = submit expected 1 but got 0
